l10n_ch_payment_slip: fix `_get_address_lines`
`payment_slip._get_address_lines` computes the address of the partner.

This is done in an onchange context whereas the changes where not isolated
and they were propagated to the real partner
via all the onchanges related to it.

We spotted the issue on the mail compose wizard
whereas the chain of onchanges was wiping all the default values.

With this change we:

* isolate the partner w/ a fake record
* additionally we cache the result per each partner
  since the function is called several times