[13.0][IMP] ebiil_paynet add qr
Implement the QR invoice functionality from SIX specifications.

Add on the ebill contract the choice of payment ESR or QR, so it can be
specific for each customer.
The XML payload has been updated. And the PDF data included improved, so
the correct PDF is generated depending of the type of payment.