l10n_ch_qr_bill: Fix _get_communications
_get_communications function defines the reference and
reference_type to be used in the generation of the QR bill.

has_qrr only checks if the reference is defined in the name
field of account.invoice. But the reference is displayed on the
report if the bank account is set up for QR bill.

This allows to generate QR XML properly even if the
name field is empty on customer invoices.