[14.0][FIX] l10n_ch_base_bank
This is a forward port of

* https://github.com/OCA/l10n-switzerland/pull/603

On an unrelated PR the test for l10n_ch_base_bank were broken.
It seems to be related to this new change introduced by

* https://github.com/odoo/odoo/pull/67599

Which test the valid of the type journal.

So this fixes the test, but I am not sure if they are still meaningfull.