Add ebill_postfinance_stock
Remove the dependency on stock in the `ebill_postfinance` module.
And add a new module `ebill_postfinance_stock` that will integrate
in the xml invoice informations related to the deliveries.

This is for Odoo implementation that sell only services and have no
stock to manage. And do not use the `stock` module.