[IMP] ebill_postfinance: use residual_amount to calculate total
The `TotalAmountDue` in the yellowBill XML now uses `invoice.amount_residual` in order to bill the remaining amount instead of the total amount in case the partner already has made partial payment. It also adds the `TotalAmountPaid` field to the XML, though I couldn't see any user-facing effects in the E-Bill portal regarding that.