ebill_postfinance: add config for xml file language
The language used to generate the xml file can affect the invoice line
description, the uom description and possibly between other.

Before this change the default language of the user click the Export
button would be used.

Now priority is on the configuration of the service, if not set the
language of the customer and then the default environment language.